Fixes for ept_intel_x64
This patch fixes a few problems with ept_intel_64::add_page and
ept_intel_x64::remove_page:

Before adding any entry that mapped a page (i.e. in the base case),
add_page cleared the m_epts if it wasn't empty. This invalidated all
the entries in that table that mapped another table.

In the recursive step (when bits > end), add_page did not
check if the entry at that level was already mapped as a page.
So if a 2M page was mapped and then a 4K page was mapped in that
same 2M region, the 2M page entry was overwritten with one
that maps a page table.

Now add_page throws an exception if the user tries to map either a page
over an entry that maps a table or a table over an entry that
already maps a page. In order to enforce the latter case, add_page now
sets bit 7 whenever a base case is reached and throws an exception if
that bit is set in a recursive case.  So with this patch, the entry
returned from add_page is cleared except for bit 7.

Since the entry returned from add_page has only bit 7 set,
root_ept_intel_x64::map_page no longer calls entry.clear()
or entry.set_entry_type() (which toggles bit 7).

Before this patch, remove_page checked if the m_epts was
empty, and if it wasn't, removed the entry from the m_epts.
Only when the m_epts was empty did the page entry get
removed (in the 'else' branch of the !m_epts.empty() check).

So if e.g. a 2M page entry and a 4K table entry were in the same
page directory, a call to remove_page for the 2M page would
see that the m_epts was not empty (because of the 4K table entry)
and would bfn::find with the 2M entry's index, which would return
NULL because that region is mapped as a page. So the 'if' branch
would exit without doing anything and the page was never removed.

With this patch, remove_page checks if the entry maps a page first
and if so clears it and returns.  If the entry doesn't map a page,
then as before it recursively removes the table entry.