x64/hve: Fix EPT test bugs
Since commit 129967715, if a user tries to map over an existing page
with a different size, an exception is thrown.

Either update the test cases that try to map over existing pages with a
different size to expect an exception or unmap the first page before
adding the second page.