vic: Fix write cr8 bug
The base handler for cr8 writes does not do the write for you. The vic's
call to cr8 in ~vic was trapping, but the hypervisor wasn't writing the
physical value, leaving the vmm susceptible to interrupt exits in
between interrupt handler destruction (i.e. in ~vic()) and guest
promotion.

Write to physical cr8 in the vic's wrcr8 handler.