Add integration test script
test_all.sh:

Constructs the list of integration vmm binaries and executes them in
sequence. After each is executed, either dmesg or 'make dump' output is
parsed and verified against expected output (of course if something else
needs to be parsed for your specific test, you can plug-in code for that too).

trap_cr8.cpp:

Now that the vic is enabled by default, the exit handlers for the cr8
integration test should return false so that the vic's handler can do
the real work.

trap_in_out.cpp:

The debug ring was being overflowed due to the amount of output printed.
In particular, the first "now in a vm" message was being truncated,
which caused the integration test to fail.

Reduce the number of print statements in order to pass the integration
test.

trap_dr7.cpp:

Return true from the test_handler to signal completion to the mov_dr
class's handler.

Signed-off-by: Connor Davis <davisc@ainfosec.com>