Pass in page for virt_lapic registers
The virtual lapic operates on its set of registers in a memory page. For
MMIO accesses, the page needs to be mapped with EPT so that the VM's
xAPIC accesses can be emulated. On the other hand, x2APIC accesses use
MSRs, so no EPT mapping is required.

Because of the different setup required, it is simpler to have the
page owned by the vic, which can do the appropriate mapping if needed.

- Replace the virt_lapic::m_registers array member with a uint32_t *
- Allocate a register page in the vic to pass to the virt_lapic constructor

Signed-off-by: Connor Davis <davisc@ainfosec.com>