Add IOAPIC intrinsics and phys/virt classes
The register space of the IOAPIC is described in Chapter 3 of
the IOAPIC specification. Its registers may be accessed with 32-bit
loads and stores offset from the base address derived from the ACPI
tables (which defaults to 0xFEC00000).

- Add phys_ioapic class to interact with the physical IOAPIC
- Add virt_ioapic class to interact with virtual IOAPICs
- Add corresponding unit tests

The current implementation provides only basic read/write capabilities
A future iteration is planned to provide more complex logic
(e.g. virtual routing) as needed.

Signed-off-by: Connor Davis <davisc@ainfosec.com>