vic: Add xapic write cache
Normally, guests only have a few different instructions that
access the xAPIC. On each write access, the exit handler
maps in the page from the guest CR3, and then disassembles the
instruction to determine what value was written.

Since the number of distinct xAPIC write instructions is small and
mapping in guest memory is so expensive, a "write cache" has been
added that associates guest rips with a virtual address in the VMM.

There is still room for more optimization around the disassembly step,
but the book keeping will be much more involved and will be put off
until it's really needed.

Signed-off-by: Connor Davis <davisc@ainfosec.com>