Improve phys_mtrr::range_list
The phys_mtrr::range_list member returns a pointer to the read-only list
of the system's memory-type ranges. An mtrr::range is created for each
range with a different type, starting from 0 and ending at the max
physical address as reported by cpuid.  The list of mtrr::ranges are
sorted by base address and represent the entire range of memory.

Signed-off-by: Connor Davis <davisc@ainfosec.com>