EFI handler changes for single-core Linux
Since the vic uses EPT to trap the xAPIC, we have to keep EPT enabled
even when unrestricted guest is off.

The EFER.LME and EFER.LMA bits have to remain in sync with
IA32E_MODE_GUEST when Linux toggles CR0.PG.

Very early in Linux boot, the VMM receives an external interrupt exit
at vector 7, which corresponds to the #NM *exception*. Usually,
exceptions cause exit reason 0, and then only if the corresponding
bit in the exception bitmap is 1 (no bits are set in the bitmap in the
current code though).

For some reason, VMX sees this as an external interrupt. This could
be because when CR0.NE is clear, vector 7 is delivered through an
external interrupt controller. But CR0.NE is required to be 1 for VM
entries on the systems so far tested.

The #NM happens on every Linux boot. There was also one external
interrupt exit at vector 0 (which is for divide error), which I'm still
trying to find an explanation for.

At any rate, Linux boots (with nosmp=1 and nox2apic=1) if the RIPs
are simply advanced in both cases, so that should do as a stop-gap.

Signed-off-by: Connor Davis <davisc@ainfosec.com>