Linux multi-core EFI
This patch enables booting SMP Linux from EFI. x2APIC is used and
seems to be working fine. Booting into SMP Linux from xAPIC gets the
cores out of boot, but eventually freezes when an interrupt-heavy task is
performed by the user (watching a cat video etc).

Much of the code here is in support for xAPIC, and thus isn't strictly
necessary for x2apic.

Signed-off-by: Connor Davis <davisc@ainfosec.com>