Remove xAPIC code
Linux enables x2APIC very early in the boot process, and so xAPIC
emulation is not needed. It is likely that any modern platform will use
x2APIC for interrupt control. For any system that doesn't support
x2APIC, interrupt emulation is disabled completely.

From the type-2 scenario, emulation is started only if the current state
of the lapic is x2APIC. If the system is in xAPIC mode, emulation is
disabled completely, and execution continues as normal.

From EFI, emulation is enabled when the guest enables x2APIC. If x2APIC
isn't supported, or the guest never switches from xAPIC, emulation does
not take place.

Signed-off-by: Connor Davis <davisc@ainfosec.com>