Optimize x2APIC emulation
xAPIC isn't really necessary, and adds alot of complexity:

  - Remove the xAPIC emulation/support code.
  - Remove the phys_lapic abstract base class. Now only phys_x2apic
    is used, with no inheritance.

When booting Linux from EFI, the INIT-SIPI-SIPI sequence is very
time sensitive. When the BSP sends the INIT IPI, the guest exits. The
BSP must ensure that the next SIPI is sent after the AP is in the
wait-for-SIPI activity state, otherwise the SIPI is dropped.

  - Signal to the BSP that the AP is "ready" in the INIT and SIPI
    handlers. Pause the BSP until the AP is ready, then procede with
    the next IPI.
  - The current signaling mechanism assumes that the APs will be
    brought up sequentially (this is how Linux does it). If a guest
    tries to bring them up in parallel, this code will likely break.

The vic and EPT memory map were being constructed. In order to give
users the choice of whether to use those modules:

  - Remove automatic enabling of EPT and VIC in the eapis

The IRR and ISR of the virt_lapic are touched on every interrupt exit.
The former implementation implemented 256-bit manip with two loops,
which is very slow.

    - Use bsr to compute the vector number in the IRR and ISR

Signed-off-by: Connor Davis <davisc@ainfosec.com>